+++ modulo3/projeto-astromatch/projeto-astromatch/src/components/Card/Card.js
@@ -3,8 +3,8 @@ import { DivPai, Header} from './CardStyled'
 import axios from "axios";
 import telamatch from '../../uteis/telamatch.svg'
 import telaprincipal from '../../uteis/telaprincipal.svg'
-import TelaMatches from "../../telas/TelaMatches";
-import TelaPrincipal from"../../telas/TelaPrincipal"
+import TelaMatches from "../../telas/TelaMatches/TelaMatches";
+import TelaPrincipal from"../../telas/TelaPrincipal/TelaPrincipal"
 
 // ----------------------------- STYLED COMPONENT
 
@@ -25,7 +25,8 @@ const Card = () => {
             setPerfilUsuario(response.data.profile)
 
         } catch (erro) {
-            console.log(erro)
+            alert("Sentimos muito. Aconteceu um erro! Volte mais tarde!")
+            console.log(erro.response.data)
         }
     }
 
@@ -35,13 +36,14 @@ const Card = () => {
             id: id,
             choice: choice
         }
-
+        
         try {
             const response = await axios.post(`${baseUrl}/choose-person`, body)
             pegarPerfil()
 
         } catch (erro) {
             console.log(erro.response.data)
+            alert("Sentimos muito. Aconteceu um erro! Volte mais tarde!")
         }
     }
 
@@ -54,13 +56,13 @@ const Card = () => {
         }
     }
 
-    //------------------------------------------------------------------------------------------------
+    //--------------------------------------------------------------------------------------
 
     return (
         <DivPai>
             <Header>
                 <img src={telaprincipal} onClick={() => setTela(true)} props = {tela}/>
-                <p>Match Perfeito</p>
+                <h1>Match Perfeito</h1>
                 <img src={telamatch} onClick={() => setTela(false)} />
             </Header>
             {trocarTela()}
